requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
mediapipe
fer
tensorflow==2.13.0
protobuf==3.20.3
